backend/requirements.txt
# Backend SalSa Market Intelligence System
# FastAPI & server
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30

# Autentikasi
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ML — SVR
scikit-learn==1.4.2

# ML — SARIMA
statsmodels==0.14.2
pmdarima==2.0.4

# Optimisasi LP
scipy==1.13.0

# Data processing
pandas==2.2.2
numpy==1.26.4

# Model persistence
joblib==1.4.2

# Export Excel
openpyxl==3.1.2

# Form data upload
python-multipart==0.0.9
